refactor: split background.js into more managable modules
