+++ src/modules/utils.js
@@ -1,0 +1,7 @@
+export function invertRecord(record) {
+  const inverted = {};
+  for (const [k, v] of Object.entries(record || {})) {
+    if (v != null) inverted[String(v)] = String(k);
+  }
+  return inverted;
+}
